Additional FDC examples (#12)
* Changed urls, made a function public, restored private key

* Breaking error fixes and cleanup

* Added minTemp weatherInsurance example

* Added weatherId weatherInsurance example

* But fix

* Camel case dir names, and removed unused data

* Added crossChainPayment FDC example

+++ contracts/PriceAnalytics.sol
@@ -2,14 +2,11 @@
 
 pragma solidity ^0.8.6;
 
-import {FtsoV2Interface} from "@flarenetwork/flare-periphery-contracts/coston2/FtsoV2Interface.sol";
-import {ContractRegistry} from "@flarenetwork/flare-periphery-contracts/coston2/ContractRegistry.sol";
+import {FtsoV2Interface} from "@flarenetwork/flare-periphery-contracts/coston/FtsoV2Interface.sol";
+import {ContractRegistry} from "@flarenetwork/flare-periphery-contracts/coston/ContractRegistry.sol";
 
 contract PriceAnalytics {
-    function convertToWei(
-        int256 price,
-        int8 decimals
-    ) public pure returns (int256) {
+    function convertToWei(int256 price, int8 decimals) public pure returns (int256) {
         int256 decimalsDiff = 18 - decimals;
         if (decimalsDiff < 0) {
             return price / int256(10 ** uint256(-decimalsDiff));
@@ -18,9 +15,7 @@ contract PriceAnalytics {
         }
     }
 
-    function provableCalculateVariance(
-        FtsoV2Interface.FeedDataWithProof[] calldata priceFeeds
-    )
+    function provableCalculateVariance(FtsoV2Interface.FeedDataWithProof[] calldata priceFeeds)
         public
         view
         returns (int256[] memory prices, int256 mean, int256 variance)
@@ -32,21 +27,14 @@ contract PriceAnalytics {
         require(prices.length > 1, "Need at least 2 prices");
         // Check correctness of prices
         for (uint256 i = 0; i < prices.length; i++) {
-            require(
-                FTSOv2.verifyFeedData(priceFeeds[i]),
-                "Price feed data is not correct"
-            );
+            require(FTSOv2.verifyFeedData(priceFeeds[i]), "Price feed data is not correct");
             if (i > 0) {
                 require(
-                    priceFeeds[i].body.votingRoundId - 1 >
-                        priceFeeds[i - 1].body.votingRoundId,
+                    priceFeeds[i].body.votingRoundId - 1 > priceFeeds[i - 1].body.votingRoundId,
                     "Price feed data is not in sequential"
                 );
             }
-            prices[i] = convertToWei(
-                int256(priceFeeds[i].body.value),
-                priceFeeds[i].body.decimals
-            );
+            prices[i] = convertToWei(int256(priceFeeds[i].body.value), priceFeeds[i].body.decimals);
         }
 
         int256 sum = 0;
@@ -54,7 +42,7 @@ contract PriceAnalytics {
             sum += prices[i];
         }
         // We incure some rounding error here, but it's not a big deal
-        mean = sum / int(prices.length);
+        mean = sum / int256(prices.length);
         for (uint256 i = 0; i < prices.length; i++) {
             if (prices[i] > mean) {
                 variance += (prices[i] - mean) * (prices[i] - mean);
